requirements.txt
streamlit
openai
python-dotenv
transformers
torch
